requirements.txt
anyio==3.6.2
autopep8==2.0.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.92.0
filelock==3.9.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
miniful==0.0.6
neo4j==5.6.0
numpy==1.24.2
orjson==3.8.7
pandas==1.5.3
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
pyFUME==0.2.25
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
scipy==1.10.1
simpful==2.10.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
